fix: use off_t/size_t/ssize_t (not long long) to match POSIX signatures
Previous declarations used long long / unsigned long long as "portable
width" but that produced signature mismatches against:

  1. Solaris 7's own stdio.h fseeko/ftello declarations (gated on
     _LARGEFILE_SOURCE) — their sig is off_t; ours was long long;
     fixincludes copied both into include-fixed/stdio.h and libgcc
     compile hit "conflicting types" errors.
  2. libsolcompat's own src/filesystem.c implementations of
     posix_fallocate/posix_fadvise (off_t in impl vs long long in
     header). ABI mismatch on ILP32 SPARC where long (off_t) is
     32-bit but long long is 64-bit — wrong register allocation.

Fix: use off_t uniformly in declarations, matching both the Solaris
prototypes and the libsolcompat implementations. off_t is visible
via <sys/types.h> which the relevant .append files now include
explicitly. ssize_t + size_t for copy_file_range match Linux's
declaration exactly.

Affected signatures:
  stdio.h.append:  fseeko, ftello
  fcntl.h.append:  posix_fallocate, posix_fadvise, fallocate
  unistd.h.append: copy_file_range
  src/process.c:   copy_file_range, fallocate (implementations)

+++ src/process.c
@@ -645,10 +645,10 @@ pthread_mutex_consistent(pthread_mutex_t *mutex)
  * using read()/write(). Our stub ensures configure probes compile.
  * ================================================================== */
 
-long long
-copy_file_range(int fd_in, long long *off_in,
-                int fd_out, long long *off_out,
-                unsigned long long len, unsigned int flags)
+ssize_t
+copy_file_range(int fd_in, off_t *off_in,
+                int fd_out, off_t *off_out,
+                size_t len, unsigned int flags)
 {
     (void)fd_in; (void)off_in; (void)fd_out; (void)off_out;
     (void)len; (void)flags;
@@ -685,7 +685,7 @@ posix_spawn_file_actions_addfchdir(posix_spawn_file_actions_t *fact, int fildes)
  * ================================================================== */
 
 int
-fallocate(int fd, int mode, long long offset, long long len)
+fallocate(int fd, int mode, off_t offset, off_t len)
 {
     (void)fd; (void)mode; (void)offset; (void)len;
     errno = ENOSYS;
